ModelsFactory/Segmentation/SEEM_workspace/git_workspace/requirements.txt
black==24.10.0
caffe2==0.8.1
cloudpickle==3.1.0
deeplearning==0.0.1
einops==0.8.0
fairscale==0.4.13
fvcore==0.1.5.post20221221
hydra-core==1.3.2
infinibatch==0.1.1
iopath==0.1.10
kornia==0.7.3
matplotlib==3.10.0
mmcv==2.2.0
mmdet==3.3.0
mpi4py==3.1.4
MultiScaleDeformableAttention==1.0
nltk==3.9.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.17.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
packaging==24.2
Pillow==11.1.0
pycocotools==2.0.8
Pygments==2.19.1
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.15.0
setuptools==72.1.0
Shapely==2.0.6
tabulate==0.9.0
termcolor==2.5.0
timm==1.0.12
torch==2.4.0+cu121
torch==2.4.0
torchvision==0.19.0
torchvision==0.19.0+cu121
tqdm==4.66.5
trainer==0.0.36
transformers==4.44.2
